MM-20225: Fix notification badge count for All Activity
When push notification setting was selected to "All Activity", we were
just sending the unread count for the channel from where the message was sent
and not for all channels. We fix this by adding a new store method: GetAnyUnreadPostCount,
which returns any unread posts from all channels.

The timer layer mocks had some additional methods which were not up to date.
They got updated as part of this.